Text correction.
original
The link between global warming and atmospheric carbon dioxide concentration is one of the keys and longest-studied links in the study of climate change.

corrected:
The link between global warming and atmospheric carbon dioxide concentration is one of the most crucial and longest-studied relationships in the study of climate change.